requirements.txt
cycler==0.10.0
imutils==0.4.6
kiwisolver==1.0.1
matplotlib==2.2.2
nltk==3.2.5
numpy==1.14.2
opencv-python==3.4.0.12
pandas==0.22.0
Pillow==5.1.0
pyparsing==2.2.0
pytesseract==0.2.0
python-dateutil==2.7.2
pytz==2018.4
six==1.11.0
tracking-url==0.0.3
xlrd==1.1.0
